refactor(core): break 2 runtime import cycles (costs, plugins)
Two genuine runtime cycles where a module imported VALUES from its own
barrel index.ts (the rest of madge's cycles are type-only `import type`
edges, harmless at runtime by design):

- costs/tools.ts imported formatCost/MODEL_PRICING/etc from ./index.js
  (which re-exports tools.ts) -> import from source modules directly.
  tools.test.ts updated to mock the source modules (was mocking the barrel).
- plugins/core-plugin.ts imported createPlugin from ./index.js -> import
  from ./registry.js.

core 57 -> 55 madge cycles; typecheck + 9,428 tests green.
